requirements.txt
absl-py==0.11.0
addict==2.4.0
ansible==2.10.3
ansible-base==2.10.3
anyio==2.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
asn1crypto==0.24.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
command-not-found==0.3
configparser==5.0.2
contextvars==2.4
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
docker-pycreds==0.4.0
dotmap==1.3.23
entrypoints==0.3
filelock==3.0.12
gast==0.2.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.35.0
h5py==3.1.0
html5lib==0.9999999
httplib2==0.9.2
idna==2.10
imageio==2.9.0
immutables==0.15
importlib-metadata==3.7.0
importlib-resources==5.1.0
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-server==1.3.0
jupyterlab==3.0.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.2.1
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
macaroonbakery==1.1.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
meld==3.18.0
mistune==0.8.4
mmcv==1.2.6
nbclassic==0.2.6
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.15.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydensecrf==1.0rc3
Pygments==2.8.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sentry-sdk==1.0.0
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==4.0.0
sniffio==1.2.0
ssh-import-id==5.7
subprocess32==3.5.4
system-service==0.3
systemd-python==234
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.57.0
traitlets==4.3.3
typing-extensions==3.7.4.3
ufw==0.35
unattended-upgrades==0.1
urllib3==1.26.3
virtualenv==20.4.2
wandb==0.10.30
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
xmltodict==0.12.0
yapf==0.30.0
youtube-dl==2018.3.14
zipp==3.4.0
